+++ GripGainsCompanion/Models/TimestampedSample.swift
@@ -2,6 +2,6 @@ import Foundation
 
 /// A force sample with device timestamp from Tindeq Progressor
 struct TimestampedSample: Equatable {
-    let weight: Float
+    let weight: Double
     let timestamp: UInt32  // microseconds from device
 }
